label-generation-demo/pyproject.toml
[tool.poetry]
name = "streamlit-demo"
version = "0.1.0"
description = ""
authors = ["Johannes Windischbauer <johannes.windischbauer@dynatrace.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
streamlit = "1.28.0"
snorkel = "^0.9.9"
scikit-learn = "1.2.0"
stqdm = "^0.0.5"
streamlit-extras = "^0.3.4"
streamlit-authenticator = "^0.2.2"
streamlit-code-editor = "^0.1.5"
streamlit-tags = "^1.2.8"
sqlalchemy = "^1.4.7"
sqlalchemy-utils = "^0.41.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
